pyproject.toml
[project]
name = "shaitan-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
